Keep failed pods

+++ source/Octopus.Tentacle/Services/Scripts/Kubernetes/KubernetesScriptServiceV1Alpha.cs
@@ -99,7 +99,8 @@ namespace Octopus.Tentacle.Services.Scripts.Kubernetes
             var response = await GetResponse(trackedPod, command.LastLogSequence, cancellationToken);
 
             //delete the pod
-            await podService.Delete(command.ScriptTicket, cancellationToken);
+            if (trackedPod.State.Phase != TrackedScriptPodPhase.Failed || !KubernetesConfig.KeepFailedScriptPods)
+                await podService.Delete(command.ScriptTicket, cancellationToken);
 
             return response;
         }
